Rename: add view in All pages
