+++ front/src/pages/AssignView.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
 
-export default function Assign() {
+export default function AssignView() {
   return <div>Assign 게시물 클릭 - 신청하기</div>;
 }
