+++ front/src/pages/NotFoundView.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
 
-export default function NotFound() {
+export default function NotFoundView() {
   return <div>Not Found</div>;
 }
