+++ front/src/pages/PostDetailView.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
 
-export default function PostDetail() {
+export default function PostDetailView() {
   return <div>게시물 클릭 (Post Detail)</div>;
 }
