+++ front/src/pages/Study/StudyView.module.css
